Refactor projection module into package
