+++ utils/capital/projection/__init__.py
@@ -1,0 +1,5 @@
+"""Capital stock projection utilities."""
+
+from .core import project_capital_stock
+
+__all__ = ["project_capital_stock"]
